+++ src/modules/shortener/usecases/update-original-url.usecase.ts
@@ -1,4 +1,5 @@
 import { Either, left, right } from '@/core/either';
+import { DatabaseError } from '@/core/errors/database.error';
 import { Injectable } from '@nestjs/common';
 import { ShortUrlRepositoryInterface } from '../repositories/short-url-repository.interface';
 import { NotFoundError } from './errors/not-found.error';
@@ -18,12 +19,19 @@ export class UpdateOriginalUrlUseCase {
     shortCode,
     originalUrl,
   }: UpdateOriginalUrlUseCaseRequest): Promise<UpdateOriginalUrlUseCaseResponse> {
-    const shortUrl = await this.shortUrlRepository.findByShortCode(shortCode);
-    if (!shortUrl) {
-      return left(new NotFoundError('ShortUrl not found'));
+    try {
+      const shortUrl = await this.shortUrlRepository.findByShortCode(shortCode);
+
+      if (!shortUrl) {
+        return left(new NotFoundError('ShortUrl not found'));
+      }
+
+      shortUrl.originalUrl = originalUrl;
+      await this.shortUrlRepository.save(shortUrl);
+
+      return right(null);
+    } catch (error) {
+      return left(new DatabaseError('Error updating original url', error));
     }
-    shortUrl.originalUrl = originalUrl;
-    await this.shortUrlRepository.save(shortUrl);
-    return right(null);
   }
 }